Trait tuxedo_core::constraint_checker::ConstraintChecker

pub trait ConstraintChecker<V>: Debug + Encode + Decode + Clone {
    type Error: Debug;
    type InherentHooks: InherentInternal<V, Self>;

    // Required methods
    fn check(
        &self,
        inputs: &[Output<V>],
        peeks: &[Output<V>],
        outputs: &[Output<V>]
    ) -> Result<TransactionPriority, Self::Error>;
    fn is_inherent(&self) -> bool;
}

A single constraint checker that a transaction can choose to call. Checks whether the input and output data from a transaction meets the codified constraints.

This full ConstraintChecker should only be implemented if the piece logic cannot be expressed with the SimpleConstraintChecker. For example, if you need to enforce the verifier is a particular type or contains a certain value. Another reason would be if you need to implement an inherent.

Additional transient information may be passed to the constraint checker by including it in the fields of the constraint checker struct itself. Information passed in this way does not come from state, nor is it stored in state.

Required Associated Types

type Error: Debug

The error type that this constraint checker may return

type InherentHooks: InherentInternal<V, Self>

Optional Associated Inherent processing logic. If this transaction type is not an inherent, use (). If it is an inherent, use Self, and implement the TuxedoInherent trait.

Required Methods

fn check( &self, inputs: &[Output<V>], peeks: &[Output<V>], outputs: &[Output<V>] ) -> Result<TransactionPriority, Self::Error>

The actual check validation logic

fn is_inherent(&self) -> bool

Tells whether this extrinsic is an inherent or not. If you return true here, you must provide the correct inherent hooks above.