Trait tuxedo_core::inherents::TuxedoInherent

pub trait TuxedoInherent<V, C: ConstraintChecker<V>>: Sized {
    type Error: Encode + IsFatalError;

    const INHERENT_IDENTIFIER: InherentIdentifier;

    // Required methods
    fn create_inherent(
        authoring_inherent_data: &InherentData,
        previous_inherent: (Transaction<V, C>, H256)
    ) -> Transaction<V, C>;
    fn check_inherent(
        importing_inherent_data: &InherentData,
        inherent: Transaction<V, C>,
        results: &mut CheckInherentsResult
    );

    // Provided method
    fn genesis_transactions() -> Vec<Transaction<V, C>> { ... }
}

Tuxedo’s controlled interface around Substrate’s concept of inherents.

This interface assumes that each inherent will appear exactly once in each block. This will be verified off-chain by nodes before block execution begins.

This interface is stricter and more structured, and therefore simpler than FRAME’s. If you need to do something more powerful (which you probably don’t) and you understand exactly how Substrate’s block authoring and Tuxedo’s piece aggregation works (which you probably don’t) you can directly implement the InherentInternal trait which is more powerful (and dangerous).

Required Associated Types

type Error: Encode + IsFatalError

Required Associated Constants

const INHERENT_IDENTIFIER: InherentIdentifier

Required Methods

fn create_inherent( authoring_inherent_data: &InherentData, previous_inherent: (Transaction<V, C>, H256) ) -> Transaction<V, C>

Create the inherent extrinsic to insert into a block that is being authored locally. The inherent data is supplied by the authoring node.

fn check_inherent( importing_inherent_data: &InherentData, inherent: Transaction<V, C>, results: &mut CheckInherentsResult )

Perform off-chain pre-execution checks on the inherent. The inherent data is supplied by the importing node. The inherent data available here is not guaranteed to be the same as what is available at authoring time.

Provided Methods

fn genesis_transactions() -> Vec<Transaction<V, C>>

Return the genesis transactions that are required for this inherent.

Object Safety

This trait is not object safe.