Trait tuxedo_core::redeemer::Redeemer

pub trait Redeemer: Debug + Encode + Decode + Clone {
    // Required method
    fn redeem(&self, simplified_tx: &[u8], witness: &[u8]) -> bool;
}

A means of checking that an output can be redeemed (aka spent). This check is made on a per-output basis and neither knows nor cares anything about the verification logic that will be applied to the transaction as a whole. Nonetheless, in order to avoid malleability, we we take the entire stripped and serialized transaction as a parameter.

Required Methods

fn redeem(&self, simplified_tx: &[u8], witness: &[u8]) -> bool