Trait tuxedo_core::Verifier

pub trait Verifier: Debug + Encode + Decode + Clone {
    type Error: Debug;

    // Required method
    fn verify(
        &self,
        input_data: &[DynamicallyTypedData],
        output_data: &[DynamicallyTypedData]
    ) -> Result<TransactionPriority, Self::Error>;
}

A single verifier that a transaction can choose to call. Verifies whether the input and output data from a transaction meets the codified constraints.

Additional transient information may be passed to the verifier by including it in the fields of the verifier struct itself. Information passed in this way does not come from state, nor is it stored in state.

Required Associated Types

type Error: Debug

The error type that this verifier may return

Required Methods

fn verify( &self, input_data: &[DynamicallyTypedData], output_data: &[DynamicallyTypedData] ) -> Result<TransactionPriority, Self::Error>

The actual verification logic