Crate tuxedo_core

This crate is the core of the Tuxedo runtime framework.

All Tuxedo runtimes will use this machinery and plug in their specific Tuxedo piece(s)

Re-exports

pub use redeemer::Redeemer;

pub use verifier::SimpleVerifier;

pub use verifier::Verifier;

Modules

dynamic_typing

Dynamic Typing utilities for Tuxedo runtimes

redeemer

A redeemer is a piece of logic that determines whether an input can be consumed in a given context. Because there are multiple reasonable ways to make this decision, we expose a trait to encapsulate the various options. Each runtime will choose to make one or more redeemers available to its users and they will be aggregated into an enum. The most common and useful redeemers are included here with Tuxedo core, but downstream developers are expected to create their own as well.

support_macros

These macros are copied from frame-support. Substrate maintainers are not open to putting them in a more sensible location. See https://github.com/paritytech/substrate/issues/13456

types

The common types that will be used across a Tuxedo runtime, and not specific to any one piece

utxo_set

These three functions comprize what I sketched out in the UtxoSet trait in element the other day For now, this is complicated enough, so I’ll just leave them here. In the future it may be wise to abstract away the utxo set. Especially if we start doing zk stuff and need the nullifiers.

verifier

A verifier is a piece of logic that determines whether a transaction as a whole is valid and should be committed. Most tuxedo pieces will provide one or more verifiers. Verifiers do not typically calculate the correct final state, but rather determine whether the proposed final state (as specified by the output set) meets the necessary constraints.

Macros

ensure

Evaluate $x:expr and if not true return Err($y:expr).

fail

Return Err of the expression: return Err($expression);.

Structs

Executive

The executive. Each runtime is encouraged to make a type alias called Executive that fills in the proper generic types.