Trait tuxedo_core::constraint_checker::ConstraintChecker

pub trait ConstraintChecker<V: Verifier>: Debug + Encode + Decode + Clone {
    type Error: Debug;

    // Required method
    fn check(
        &self,
        inputs: &[Output<V>],
        outputs: &[Output<V>]
    ) -> Result<TransactionPriority, Self::Error>;
}

A single constraint checker that a transaction can choose to call. Checks whether the input and output data from a transaction meets the codified constraints.

This full ConstraintChecker should only be used if there is more that a piece wants to do such as check the verifier information in some unique way.

Additional transient information may be passed to the constraint checker by including it in the fields of the constraint checker struct itself. Information passed in this way does not come from state, nor is it stored in state.

Required Associated Types

type Error: Debug

the error type that this constraint checker may return

Required Methods

fn check( &self, inputs: &[Output<V>], outputs: &[Output<V>] ) -> Result<TransactionPriority, Self::Error>

The actual check validation logic