Module tuxedo_core::verifier

A verifier is a piece of logic that determines whether an input can be consumed in a given context. Because there are multiple reasonable ways to make this decision, we expose a trait to encapsulate the various options. Each runtime will choose to make one or more verifiers available to its users and they will be aggregated into an enum. The most common and useful verifiers are included here with Tuxedo core, but downstream developers are expected to create their own as well.

Structs

SigCheck

A typical verifier that checks an sr25519 signature

SignatureAndIndex

Combination of a signature plus and index so that the signer can specify which index this signature pertains too of the available signatories for a ThresholdMultiSignature

TestVerifier

A testing verifier that passes or depending on the enclosed boolean value.

ThresholdMultiSignature

A Threshold multisignature. Some number of member signatories collectively own inputs guarded by this verifier. A valid redeemer must supply valid signatures by at least threshold of the signatories. If the threshold is greater than the number of signatories the input can never be consumed.

UpForGrabs

A simple verifier that allows anyone to consume an output at any time

Traits

Verifier

A means of checking that an output can be verified (aka spent). This check is made on a per-output basis and neither knows nor cares anything about the validation logic that will be applied to the transaction as a whole. Nonetheless, in order to avoid malleability, we we take the entire stripped and serialized transaction as a parameter.