Trait tuxedo_core::verifier::Verifier

pub trait Verifier: Debug + Encode + Decode + Clone {
    // Required method
    fn verify(&self, simplified_tx: &[u8], redeemer: &[u8]) -> bool;
}

A means of checking that an output can be verified (aka spent). This check is made on a per-output basis and neither knows nor cares anything about the validation logic that will be applied to the transaction as a whole. Nonetheless, in order to avoid malleability, we we take the entire stripped and serialized transaction as a parameter.

Required Methods

fn verify(&self, simplified_tx: &[u8], redeemer: &[u8]) -> bool